backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dj-database-url==2.3.0
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gitdb==4.0.12
GitPython==3.1.41
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
oracledb==3.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
setuptools==75.6.0
smmap==5.0.2
sqlparse==0.5.3
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.4.0
